refactor: FriendshipsRepositoryのRequestクラスとResponseクラスを分離
