+++ lib/models/repositories/friendships/friendships_api_request.dart
@@ -1,0 +1,11 @@
+class FriendshipsApiRequest {
+  final String userIds;
+
+  FriendshipsApiRequest({
+    required this.userIds,
+  });
+
+  Map<String, String> toParams() => {
+    "user_ids": this.userIds,
+  };
+}
