+++ lib/models/repositories/friendships/friendships_repository.dart
@@ -1,8 +1,9 @@
 import 'dart:convert';
 import 'dart:io';
-import 'package:event_follow/models/entities/user.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:http/http.dart' as http;
+import 'friendships_api_request.dart';
+import 'friendships_api_response.dart';
 
 final friendshipsRepositoryProvider = Provider.autoDispose.family<FriendshipsRepository, dynamic>(
         (ref, getOrGenerateIdToken) => FriendshipsRepository(getOrGenerateIdToken: getOrGenerateIdToken));
@@ -12,7 +13,7 @@ class FriendshipsRepository {
 
   FriendshipsRepository({required this.getOrGenerateIdToken});
 
-  Future<FriendshipsApiResults> requestFriendshipsApi(
+  Future<FriendshipsApiResponse> requestFriendshipsApi(
       {required FriendshipsApiRequest request}) async {
     final url = Uri.https("event-follow-front.herokuapp.com",
         "/api/friendships", request.toParams());
@@ -25,32 +26,9 @@ class FriendshipsRepository {
       },
     );
 
-    return FriendshipsApiResults.fromJson(json.decode(response.body));
+    return FriendshipsApiResponse.fromJson(json.decode(response.body));
   }
 }
 
-class FriendshipsApiRequest {
-  final String userIds;
 
-  FriendshipsApiRequest({
-    required this.userIds,
-  });
-
-  Map<String, String> toParams() => {
-    "user_ids": this.userIds,
-  };
-}
-
-class FriendshipsApiResults {
-  FriendshipsApiResults({
-    required this.friends,
-  });
-
-  List<User> friends;
-
-  factory FriendshipsApiResults.fromJson(List<dynamic> json) =>
-      FriendshipsApiResults(
-        friends: List<User>.from(json.map((x) => User.fromJson(x))),
-      );
-}
