+++ lib/pages/events_pages/friends_footer.dart
@@ -3,6 +3,7 @@ import 'package:event_follow/models/controllers/friendships_controller/friendshi
 import 'package:event_follow/models/entities/event.dart';
 import 'package:event_follow/models/entities/event_extra.dart';
 import 'package:event_follow/models/models.dart';
+import 'package:event_follow/models/repositories/friendships/friendships_api_request.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
